Create Book Array Based on Category
The first functionality of this application is to get and display
the books of a specific category. This commit covers the 'getting'
part.

Once the user filled the category field and clicked the search button
an API call is made with the specific category the user looked for.
With the returned data we create and return a new array that contains
only relevant data that will be displayed.