Add Display Book Previews Functionality
Once all the books of a specific category are retrieved from the API
they must be displayed.

We send the new array of book objects to a function that for each book
in the array display it with its cover, title and author. A click event
is added on each book because it will allow to show the specification
when the user click on it.