Get Book Description
Once all the previews are displayed the user can click on them
to display their specifics and a description of the book.

To do that we get the book number in the array with a global event
listener, if the target contains specific classes it means that
the user want to read the specific of the closest book in the list.
With that number we retrieve the clicked book in the array so
we can use its key to get the description from API. Book and its
description are returned.