Add Display Book Specific Functionality
Once we have retrieved the description of the clicked book in preview
list we need to display its specifics.

With a function that gets as parameters the clicked book and its description
we remove the hidden class from the modal that contains the elements
of the book and fill the fields with the details that must be displayed
(cover, title, author and description).