Add Functionality to Close Book Specifics
Clicking on the X button or on the overlay outside the container of
book specifics must add hidden class to container.

We can do it using the same function we used to show the container
but converting from add class to toggle class (when container contains
hidden then remove hidden (close modal), when container does not
contains hidden, add hidden (open modal)).