Error Handling Searching Category
There may be type of errors when the user search for a category:
- Input is empty
- Internet is slow, the request requires more time than timeout
- Category does not exist
All these errors must be displayed on the UI.

In the same function where we get the books for category we handle all
those errors. Axios gives useful conditions and details that we can
use as condition for each type of problem. The error is returned as
value of the array.