Error Handling Book Description Not Available
Some book does not have a description available from the API, if
the user click on a book without a description must be notified that
that book has no description available.

Since this is the only error that we need to handle in this case, we
check if the response object coming from the API has the property
'description', if not it will return an error that it's prepended
to the book when the user click on the book without a description.